After a brainstorming session with @anuchandy, we decided that if we are going to decouple Android classes like Context and Configuration to make it easier to create some unit tests it would be good to reduce the surface area of our public API involved (InformationProvider interfaces). So instead of having a whole 'implementation' package with a bunch of public interfaces and classes, we can consolidate everything into one interface and its implementation and move it into the 'interceptors' package while keeping the implementation package-private and the interface public with a getDefault() method.
