--- sdk/core/azure-core/src/main/java/com/azure/android/core/provider/LocaleInformationProvider.java
@@ -1,20 +1,0 @@
-package com.azure.android.core.provider;
-
-/**
- * Interface for a provider that contains locale information.
- */
-public interface LocaleInformationProvider {
-    /**
-     * Default system language.
-     *
-     * @return The default system language.
-     */
-    String getDefaultSystemLanguage();
-
-    /**
-     * System region.
-     *
-     * @return The system region.
-     */
-    String getSystemRegion();
-}
