--- sdk/core/azure-core/src/main/java/com/azure/android/core/provider/PlatformInformationProvider.java
@@ -1,27 +1,0 @@
-package com.azure.android.core.provider;
-
-/**
- * Interface for a provider that contains platform information.
- */
-public interface PlatformInformationProvider {
-    /**
-     * Device manufacturer.
-     *
-     * @return The device manufacturer.
-     */
-    String getManufacturer();
-
-    /**
-     * Device model.
-     *
-     * @return The device model.
-     */
-    String getModel();
-
-    /**
-     * OS version.
-     *
-     * @return The device's OS version.
-     */
-    int getOsVersion();
-}
