--- sdk/core/azure-core/src/main/java/com/azure/android/core/provider/implementation/AndroidLocaleInformationProvider.java
@@ -1,54 +1,0 @@
-package com.azure.android.core.provider.implementation;
-
-import android.content.res.Configuration;
-
-import androidx.annotation.NonNull;
-
-import com.azure.android.core.provider.LocaleInformationProvider;
-
-/**
- * Provider that contains system locale information extracted using given {@link Configuration} object. The former
- * can be obtained by calling {@code Resources.getSystem().getConfiguration()}.
- */
-public class AndroidLocaleInformationProvider implements LocaleInformationProvider {
-    private final Configuration configuration;
-    private String language;
-    private String systemRegion;
-
-    /**
-     * Constructor that takes a system {@link Configuration} to extract data from.
-     *
-     * @param configuration The system configuration.
-     */
-    public AndroidLocaleInformationProvider(@NonNull Configuration configuration) {
-        this.configuration = configuration;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getDefaultSystemLanguage() {
-        if (language == null) {
-            // Using this instead of Configuration.getLocales() because it's not supported in anything less than
-            // Android L24
-            language = configuration.locale.getLanguage();
-        }
-
-        return language;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getSystemRegion() {
-        if (systemRegion == null) {
-            // Using this instead of Configuration.getLocales() because it's not supported in anything less than
-            // Android L24
-            systemRegion = configuration.locale.getCountry();
-        }
-
-        return systemRegion;
-    }
-}
