--- sdk/core/azure-core/src/main/java/com/azure/android/core/provider/implementation/AndroidPlatformInformationProvider.java
@@ -1,34 +1,0 @@
-package com.azure.android.core.provider.implementation;
-
-import android.os.Build;
-
-import com.azure.android.core.provider.PlatformInformationProvider;
-
-/**
- * Provider that contains platform information extracted from the {@link Build} class.
- */
-class AndroidPlatformInformationProvider implements PlatformInformationProvider {
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getManufacturer() {
-        return Build.MANUFACTURER;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getModel() {
-        return Build.MODEL;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public int getOsVersion() {
-        return Build.VERSION.SDK_INT;
-    }
-}
